requirements.txt
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
psycopg2_binary==2.9.7
flask
plotly
paramiko<3.0.0
sshtunnel